requirements.txt
Flask
numpy
Pillow
pydub
audioop-lts
